Renamed Open Dialog files to Select Dialog
