+++ src/renderer/components/project-dialogs/select-project-tab.component.scss
